requirements.txt
# ============ Core Dependencies ============
numpy>=2.0.0,<2.3.0
pandas>=2.0.0
pyyaml>=6.0
scipy>=1.10.0
scikit-learn>=1.2.0,<1.3.0

# ============ Optimization ============
pymoo>=0.6.0
pygmo>=2.19.0

# ============ LLM ============
openai>=1.0.0
transformers>=4.30.0
torch>=2.0.0

# ============ Utilities ============
requests>=2.31.0
matplotlib>=3.7.0
tqdm>=4.65.0

# ============ Chemistry (Optional) ============
pytdc>=0.3.0
# rdkit-pypi
